Revert "Feature/add get cursos controller"
