+++ src/controllers/LoginController.js
@@ -1,6 +1,5 @@
-const { compare } = require("bcryptjs");
-const Aluno = require('../models/Aluno');
-const { sign } = require('jsonwebtoken');
+const Aluno = require('../models/Aluno')
+const { sign } = require('jsonwebtoken')
 
 class LoginController {
 
@@ -19,19 +18,12 @@ class LoginController {
 
             //Busca o aluno pelo email
             const aluno = await Aluno.findOne({
-                where: { email: email }
+                where: { email: email, senha: senha }
             });
 
             //Verifica se o aluno existe
             if (!aluno) {
-                return res.status(404).json({ message: 'Nenhum aluno corresponde ao email fornecido!' });
-            }
-
-            //Compara senhas com as senhas hasheadas do bcryptjs
-            const senhaValida = await compare(senha, aluno.senha);
-
-            if(senhaValida === false) {
-                return res.status(403).json({message: 'Senha inválida!'});
+                return res.status(404).json({ message: 'Nenhum aluno corresponde ao email e senha fornecidos!' });
             }
 
             //Verifica autenticação
