+++ src/models/Aluno.js
@@ -1,6 +1,5 @@
 const { DataTypes } = require('sequelize');
 const { connection } = require('../database/connection');
-const { hash } = require("bcryptjs");
 
 const Aluno = connection.define('alunos', {
     nome: {
@@ -18,12 +17,6 @@ const Aluno = connection.define('alunos', {
     data_nascimento: {
         type: DataTypes.DATE
     }
-});
-
-//HOOK
-Aluno.beforeSave(async (user) => {
-    user.senha = await hash(user.senha, 8)
-    return user
 });
 
 module.exports = Aluno;
